requirements.txt
torch>=2.0.0
numpy>=1.21.0
h5py>=3.7.0
Pillow>=9.0.0
scipy>=1.7.0 